patchfns: Extend format of series file
Support whitespace before patch file name
Interpret lines that begin with +- as comments (in fact, they are "guards")
Reuse leading space from previous entry when inserting a new entry

This effectively allows using series.conf directly as a series file. The
benefit is that the former doesn't need to be updated after using `quilt
import` and other functions that change the series file.